* merged from trunk r18206:18273.
git-svn-id: http://svn.ruby-lang.org/repos/ruby/branches/mvm@18274 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ test/win32ole/test_win32ole_event.rb
@@ -6,6 +6,8 @@ require 'test/unit'
 
 if defined?(WIN32OLE_EVENT)
   class TestWIN32OLE_EVENT < Test::Unit::TestCase
+    module IE
+    end
     def create_temp_html
       fso = WIN32OLE.new('Scripting.FileSystemObject')
       dummy_file = fso.GetTempName + ".html"
@@ -17,9 +19,25 @@ if defined?(WIN32OLE_EVENT)
       dummy_path
     end
 
+    def message_loop
+      WIN32OLE_EVENT.message_loop
+      sleep 0.1
+    end
+
+    def wait_ie
+      while @ie.readyState != IE::READYSTATE_COMPLETE
+        message_loop
+      end
+    end
+
     def setup
+      WIN32OLE_EVENT.message_loop
       @ie = WIN32OLE.new("InternetExplorer.Application")
+      if !defined?(IE::READYSTATE_COMPLETE)
+        WIN32OLE.const_load(@ie, IE)
+      end
       @ie.visible = true
+      message_loop
       @event = ""
       @event2 = ""
       @event3 = ""
@@ -43,8 +61,7 @@ if defined?(WIN32OLE_EVENT)
       while @ie.busy
         WIN32OLE_EVENT.new(@ie)
         GC.start  
-        WIN32OLE_EVENT.message_loop
-        sleep 0.1
+        message_loop
       end
       assert_match(/BeforeNavigate/, @event)
       assert_match(/NavigateComplete/, @event)
@@ -54,11 +71,7 @@ if defined?(WIN32OLE_EVENT)
       ev = WIN32OLE_EVENT.new(@ie, 'DWebBrowserEvents')
       ev.on_event {|*args| default_handler(*args)}
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        WIN32OLE_EVENT.new(@ie, 'DWebBrowserEvents') 
-        GC.start  
-        sleep 0.1
-      end
+      wait_ie
       assert_match(/BeforeNavigate/, @event)
       assert_match(/NavigateComplete/, @event)
     end
@@ -68,10 +81,7 @@ if defined?(WIN32OLE_EVENT)
       ev.on_event('BeforeNavigate') {|*args| handler1}
       ev.on_event('BeforeNavigate') {|*args| handler2}
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        WIN32OLE_EVENT.message_loop
-        sleep 0.1
-      end
+      wait_ie
       assert_equal("handler2", @event2)
     end
 
@@ -80,10 +90,7 @@ if defined?(WIN32OLE_EVENT)
       ev.on_event {|*args| handler1}
       ev.on_event {|*args| handler2}
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        WIN32OLE_EVENT.message_loop
-        sleep 0.1
-      end
+      wait_ie
       assert_equal("handler2", @event2)
     end
 
@@ -93,10 +100,7 @@ if defined?(WIN32OLE_EVENT)
       ev.on_event{|*args| handler2}
       ev.on_event('NavigateComplete'){|*args| handler3(*args)}
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        WIN32OLE_EVENT.message_loop
-        sleep 0.1
-      end
+      wait_ie
       assert(@event3!="")
       assert("handler2", @event2)
     end
@@ -106,10 +110,7 @@ if defined?(WIN32OLE_EVENT)
       ev.on_event {|*args| default_handler(*args)}
       ev.on_event('NavigateComplete'){|*args| handler3(*args)}
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        WIN32OLE_EVENT.message_loop
-        sleep 0.1
-      end
+      wait_ie
       assert_match(/BeforeNavigate/, @event)
       assert(/NavigateComplete/ !~ @event)
       assert(@event!="")
@@ -119,18 +120,12 @@ if defined?(WIN32OLE_EVENT)
       ev = WIN32OLE_EVENT.new(@ie, 'DWebBrowserEvents')
       ev.on_event {|*args| default_handler(*args)}
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        WIN32OLE_EVENT.message_loop
-        sleep 0.1
-      end
+      wait_ie
       assert_match(/BeforeNavigate/, @event)
       ev.unadvise
       @event = ""
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        WIN32OLE_EVENT.message_loop
-        sleep 0.1
-      end
+      wait_ie
       assert_equal("", @event);
       assert_raise(WIN32OLERuntimeError) {
         ev.on_event {|*args| default_handler(*args)}
@@ -157,10 +152,7 @@ if defined?(WIN32OLE_EVENT)
       }
       bl = @ie.locationURL
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        sleep 0.1
-        WIN32OLE_EVENT.message_loop
-      end
+      wait_ie
       assert_equal(bl, @ie.locationURL)
     end
 
@@ -171,10 +163,7 @@ if defined?(WIN32OLE_EVENT)
       }
       bl = @ie.locationURL
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        sleep 0.1
-        WIN32OLE_EVENT.message_loop
-      end
+      wait_ie
       assert_equal(bl, @ie.locationURL)
     end
 
@@ -185,10 +174,7 @@ if defined?(WIN32OLE_EVENT)
       }
       bl = @ie.locationURL
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        sleep 0.1
-        WIN32OLE_EVENT.message_loop
-      end
+      wait_ie
       assert_equal(bl, @ie.locationURL)
     end
 
@@ -199,10 +185,7 @@ if defined?(WIN32OLE_EVENT)
       }
       bl = @ie.locationURL
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        sleep 0.1
-        WIN32OLE_EVENT.message_loop
-      end
+      wait_ie
       assert_equal(bl, @ie.locationURL)
     end
     
@@ -213,10 +196,7 @@ if defined?(WIN32OLE_EVENT)
       }
       bl = @ie.locationURL
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        sleep 0.1
-        WIN32OLE_EVENT.message_loop
-      end
+      wait_ie
       assert_equal(bl, @ie.locationURL)
     end
 
@@ -227,11 +207,26 @@ if defined?(WIN32OLE_EVENT)
       }
       bl = @ie.locationURL
       @ie.navigate("file:///#{@f}")
-      while @ie.busy
-        sleep 0.1
-        WIN32OLE_EVENT.message_loop
-      end
+      wait_ie
       assert_equal(bl, @ie.locationURL)
+    end
+
+    def test_off_event
+      ev = WIN32OLE_EVENT.new(@ie)
+      ev.on_event{handler1}
+      ev.off_event
+      @ie.navigate("file:///#{@f}")
+      wait_ie
+      assert_equal("", @event2)
+    end
+
+    def test_off_event_arg
+      ev = WIN32OLE_EVENT.new(@ie)
+      ev.on_event('BeforeNavigate2'){handler1}
+      ev.off_event('BeforeNavigate2')
+      @ie.navigate("file:///#{@f}")
+      wait_ie
+      assert_equal("", @event2)
     end
 
     def handler1
@@ -248,21 +243,72 @@ if defined?(WIN32OLE_EVENT)
 
     def teardown
       @ie.quit
-      WIN32OLE_EVENT.message_loop
+      message_loop
       @ie = nil
-      WIN32OLE_EVENT.message_loop
-      sleep 0.1
+      i = 0
       begin 
-        File.unlink(@f)
+        i += 1
+        File.unlink(@f) if i < 10
       rescue Errno::EACCES
-        WIN32OLE_EVENT.message_loop
-        sleep 0.1
-        File.unlink(@f)
+        message_loop
+        retry
       end
-
+      message_loop
       GC.start
-      WIN32OLE_EVENT.message_loop
-      sleep 0.1
+      message_loop
     end
+
+    class Handler1
+      attr_reader :val1, :val2, :val3, :val4
+      def initialize
+        @val1 = nil
+        @val2 = nil
+        @val3 = nil
+        @val4 = nil
+      end
+      def onStatusTextChange(t)
+        @val1 = t
+      end
+      def onProgressChange(p, pmax)
+        @val2 = p
+        @val3 = pmax
+      end
+      def onPropertyChange(p)
+        @val4 = p
+      end
+    end
+
+    class Handler2
+      attr_reader :ev
+      def initialize
+        @ev = ""
+      end
+      def method_missing(ev, *arg)
+        @ev += ev
+      end
+    end
+
+    def test_handler1
+      ev = WIN32OLE_EVENT.new(@ie)
+      h1 = Handler1.new
+      ev.handler = h1
+      @ie.navigate("file:///#{@f}")
+      wait_ie
+      assert(h1.val1)
+      assert_equal(h1.val1, ev.handler.val1)
+      assert(h1.val2)
+      assert(h1.val3)
+      assert(h1.val4)
+    end
+
+    def test_handler2
+      ev = WIN32OLE_EVENT.new(@ie)
+      h2 = Handler2.new
+      ev.handler = h2
+      @ie.navigate("file:///#{@f}")
+      wait_ie
+      assert(h2.ev != "")
+    end
+
   end
 end
